Se realisaron los cambios y se eliminaron compo
