--- src/app/pages/Dialog/dialog-change-pass/dialog-change-pass.component.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head> </head>
-  <body style="background-color: white">
-    <div class="contenedor">
-      <button
-        class="elemento"
-        mat-mini-fab
-        color="secondary"
-        (click)="closeDialog()"
-      >
-        <mat-icon>close</mat-icon>
-      </button>
-    </div>
-    <h3 class="text-center text">¡Mensaje de recuperación </h3>
-    
-    <div class="row" style=" margin-top: 30px;">
-        <div class="col-2">  <mat-icon style="font-size: 50px">
-            <span class="material-symbols-outlined">
-                check_small
-            </span>
-        </mat-icon></div>
-        <div class="col-10 text text-Justified  d-flex align-items-center " > <h4>
-            Se ha guardado correctamente tu nueva contraseña.
-        </h4></div>
-    </div>
-    
-  </body>
-</html>
