--- src/app/pages/Dialog/dialog-change-pass/dialog-change-pass.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DialogChangePassComponent } from './dialog-change-pass.component';
-
-describe('DialogChangePassComponent', () => {
-  let component: DialogChangePassComponent;
-  let fixture: ComponentFixture<DialogChangePassComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ DialogChangePassComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(DialogChangePassComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
