--- src/app/pages/Dialog/dialog-recover-pass/dialog-recover-pass.component.scss
@@ -1,42 +1,0 @@
-.contenInputCode {
-  display: flex;
-  justify-content: center;
-  margin-top: 40px;
-  margin: 20px;
-}
-
-.inputSucetfull {
-  width: 100%;
-  height: 80px;
-  text-align: center;
-  border-radius: 10px;
-  border: 2pt solid #0071bc;
-  border-radius: 5px;
-  font-size: 50pt;
-}
-
-.contenedor {
-  position: relative;
-}
-
-.elemento {
-  position: absolute;
-  right: 0;
-}
-.inputError {
-  width: 100%;
-  height: 80px;
-  text-align: center;
-  border-radius: 10px;
-  border: 2pt solid red;
-  border-radius: 5px;
-  font-size: 50pt;
-}
-.menssageError{
-  margin-top: 5px;
-  color: red;
-  font-size: 12px;
-}
-
-
-
