--- src/app/pages/Dialog/dialog-signup-error/dialog-signup-error.component.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head> </head>
-  <body style="background-color: white">
-    <div class="contenedor">
-      <button
-        class="elemento"
-        mat-mini-fab
-        color="secondary"
-        (click)="closeDialog()"
-      >
-        <mat-icon>close</mat-icon>
-      </button>
-    </div>
-    <h3 class="text-center text">¡Mensaje de registro!</h3>
-    
-    <div class="row" style=" margin-top: 30px;">
-        <div class="col-2">  <mat-icon style="font-size: 50px; color: red;">
-            <span class="material-symbols-outlined">
-                close
-                </span>
-        </mat-icon></div>
-        <div class="col-10 text text-Justified  d-flex align-items-center " > <h5>
-            {{data}}
-        </h5></div>
-    </div>
-    
-  </body>
-</html>
