--- src/app/pages/Dialog/dialog-signup-error/dialog-signup-error.component.scss
@@ -1,8 +1,0 @@
-.contenedor {
-    position: relative;
-  }
-  
-  .elemento {
-    position: absolute;
-    right: 0;
-  }
