+++ src/app/pages/auth/recover-password/recover-password.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from "@angular/core";
 import { FormBuilder, FormGroup, Validators } from "@angular/forms";
 import { MatDialog, MatDialogConfig } from "@angular/material/dialog";
 import { ActivatedRoute, Router } from "@angular/router";
-import { DialogRecoverPassComponent } from "../../Dialog/dialog-recover-pass/dialog-recover-pass.component";
+
 
 @Component({
   selector: "app-recover-password",
@@ -45,17 +45,9 @@ export class RecoverPasswordComponent implements OnInit {
     this.submitted = true;
     // stop here if form is invalid
     if (this.authForm.valid) {
-      this.openDialog();
+      
     }
   }
 
-  openDialog() {
-    const dialogConfig: MatDialogConfig<any> = {
-      maxWidth: "100vw",
-      maxHeight: "100vh",
-      panelClass: "full-screen-dialog",
-    };
-    dialogConfig.disableClose = true;
-    this.matDialog.open(DialogRecoverPassComponent,dialogConfig)
-  }
+
 }
